build: first attempt at compiling for Win64
Notes:
- build only, no real support yet
- some code was copied from Obsoletium project

+++ src/public/vgui_controls/RichText.h
@@ -154,7 +154,7 @@ protected:
 	virtual void OnSetFocus();
 
 	// clickable url handling
-	int ParseTextStringForUrls(const char *text, int startPos, char *pchURLText, int cchURLText, char *pchURL, int cchURL, bool &clickable);
+	int ParseTextStringForUrls(const char *text, intp startPos, char *pchURLText, intp cchURLText, char *pchURL, intp cchURL, bool &clickable);
 	virtual void OnTextClicked(const wchar_t *text);
 
 #ifdef DBGFLAG_VALIDATE
